update
git-svn-id: http://svn.tudos.org/repos/tudos/trunk@103 a704ac0b-3a55-4d43-a2a9-7be6f07c34fb

+++ l4/pkg/dope/lib/dope/common/main.c
@@ -75,7 +75,7 @@ void dope_bindf(long id, const char *varfmt, const char *event_type,
 void dope_process_event(long id) {
 	dope_event *e;
 	char *bindarg, *s;
-	u32 num1 = 0, num2 = 0;
+	unsigned long num1 = 0, num2 = 0;
 
 	dopelib_wait_event(id, &e, &bindarg);
